OLH-1619: Consolidate type definitions
