--- src/common/oidc-token.models.ts
@@ -1,7 +1,0 @@
-export interface TokenResponse {
-  access_token: string;
-  refresh_token: string;
-  token_type: string;
-  expires_in: number;
-  id_token: string;
-}
