--- src/common/oidc-userinfo-models.ts
@@ -1,8 +1,0 @@
-export interface UserInfo {
-  sub: string;
-  email: string;
-  email_verified: boolean;
-  phone_number: string;
-  phone_number_verified: boolean;
-  updated_at: string;
-}
